Update delete functionality to support delete by ID

+++ src/main/java/cms/logic/parser/DeleteCommandParser.java
@@ -1,12 +1,13 @@
 package cms.logic.parser;
 
-import static cms.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-
 import java.util.List;
+import java.util.stream.Collectors;
 
 import cms.commons.core.index.Index;
+import cms.logic.Messages;
 import cms.logic.commands.DeleteCommand;
 import cms.logic.parser.exceptions.ParseException;
+import cms.model.person.NusId;
 
 /**
  * Parses input arguments and creates a new DeleteCommand object
@@ -21,12 +22,33 @@ public class DeleteCommandParser implements Parser<DeleteCommand> {
      */
     public DeleteCommand parse(String args) throws ParseException {
         try {
-            List<Index> indexes = ParserUtil.parseIndexes(args);
+            String trimmedArgs = args.trim();
+            if (trimmedArgs.startsWith("id/")) {
+                List<NusId> nusIds = ParserUtil.parseNusIds(parseNusIdTokens(trimmedArgs));
+                return DeleteCommand.byNusIds(nusIds);
+            }
+
+            List<Index> indexes = ParserUtil.parseIndexes(trimmedArgs);
             return new DeleteCommand(indexes);
         } catch (ParseException pe) {
             throw new ParseException(
-                    String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_USAGE), pe);
+                    String.format(Messages.MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_USAGE), pe);
         }
+    }
+
+    private List<String> parseNusIdTokens(String trimmedArgs) throws ParseException {
+        String[] tokens = trimmedArgs.split("\\s+");
+        if (tokens.length == 0) {
+            throw new ParseException(NusId.MESSAGE_CONSTRAINTS);
+        }
+
+        return List.of(tokens).stream()
+                .map(this::stripNusIdPrefix)
+                .collect(Collectors.toList());
+    }
+
+    private String stripNusIdPrefix(String token) {
+        return token.startsWith("id/") ? token.substring(3) : token;
     }
 
 }
